Move rotor velocity tests to individual module.
